[GLOBAL] {Removed all warning messages} Read new comment lines CAREFULLY!

+++ centurion-source/source/engine/classes/unit.cpp
@@ -118,7 +118,7 @@ void Unit::render(bool picking, int clickID, bool not_placeable)
 			hitbox.rectangle.render(
 				this->IsSelected() ? glm::vec4(255.0f, 0.0f, 255.0f, 1.0f) : glm::vec4(255.0f, 242.0f, 0.0f, 1.0f),
 				0,
-				position3D.x,
+				(int)position3D.x,
 				position3D.y + unitData.yOffset,
 				0,
 				0
